findline tune

+++ src/lib/lfindline.cxx
@@ -587,7 +587,7 @@ TESTS:
 
     dx = abs(x1 - x2);
     dy = abs(y1 - y2);
-    dist= sqrt(dx*dx+dy*dy); //calcul la distance entre les deux points
+    dist= sqrt(double(dx*dx+dy*dy)); //calcul la distance entre les deux points
 
     incx = 1;
     if (x2 < x1)
